Refactor C API to avoid using unwrap. (#12)
This removes another source of possible crashes from the rust code.

I also took this opportunity to try and test the C-api. I also
discovered rust "integration tests" as part of this and chose
to move some of the (new) c_api tests to an integration test
as well as the existing rust api (updater.rs) tests.