refactor: remove client_id
We added client_id back when we had a different billing model
where we would bill customers based on number of unique devices
accepting an install in a given month (similar to what expo.dev
advertised as their pricing model at the time).

We moved to a differnet pricing model and never used client_id
so removing it.